+++ frontend/src/pages/RoomSelect.jsx
@@ -1,7 +1,7 @@
 import React, {useState} from 'react'
 import { useHistory, useLocation } from 'react-router-dom';
 import { GetRoom } from "../http/Rooms";
-import './RoomSelect.css';
+import './css/RoomSelect.css';
 import {Oval} from "@agney/react-loading";
